Use require.main.require 'hubot' instead of require 'hubot'
Because we have hubot in our devDependencies, if these dependencies are
installed, `require 'hubot'` will pick up the local version of hubot
instead of the one that the robot is running from. This means that our
emitted `TextMessage`s won't trigger any `TextListener`s installed by
scripts (since that uses `instanceof`).

As a workaround, use `require.main.require 'hubot'` instead, which
requires hubot from the context of the main module of the program. This
should produce the same module that the robot is running from (barring
unusual edge cases). If that fails, fall back to `require 'hubot'`,
to handle cases such as running the tests.