Avoid our devDependencies when requiring 'hubot'
Because we have hubot in our devDependencies, if these dependencies are
installed, `require 'hubot'` will pick up the local version of hubot
instead of the one that the robot is running from. This means that our
emitted `TextMessage`s won't trigger any `TextListener`s installed by
scripts (since that uses `instanceof`).

Implement a workaround that avoids loading 'hubot' from our local
node_modules folder if possible. It first tries finding 'hubot' without
our node_modules, and if that fails, it walks up the module parent tree
until it finds a module outside our directory tree and tries again from
that spot. This means it works correctly even if hubot-slack is
installed with `npm link`. It falls back to a normal `require 'hubot'`
if all else fails, which allows the test suite to work.