Allow sending DMs to arbitrary users
This change uses the new callbacks available in node-slack-client. If
robot.send is called like:

robot.send {room: 'someuser'}, 'message'

If someuser is a user and there is no existing DM, it will use openDM to
establish a new private message with the user and send the message in the
callback.