Copy edit for README
The current language around removing the old hubot integration implies that you can actually delete the integration, which is confusing. I went searching all over for a way to "delete" or "remove" him, since all I could find in the UI was an option to "disable". Eventually I discovered #147 which clarifies that you cannot actually delete or remove old integrations, only disable. So updating the language in the README here would be helpful. :sparkles: