Trigger enter/ leave actions by presence change
The current adapter implementation triggers robot.enter / robot.leave events when someone joins/ leaves a channel or
group. However, most Hubot scripts use these callbacks registers with the much more frequent Campfire/ XMPP/ etc
presence change events in mind (e.g. when someone logs on/ logs off). Since joining a slack channel is something you
tend to do only once, or at most once in a while, the callbacks are not so very useful at present for the slack adapter.

One downside to the new implementation is that because a presence change event is generic, not specific to a (Hubot)
room/ (Slack) channel, there is no room in the envelope passed to your enter/ leave callback. I adapted our own
scripts to use robot.messageRoom rather than envelope.reply, but am open to suggestions on how to improve this, perhaps
by looking at the feasibility on sending a enter/ leave notice to each channel the user in question is in.