Return rawText and rawMessage objects
Prior to 900402ea254b0fdf5a819fd3073325c68f1692a9, hubot-slack returned
a message object which included the raw text from the Slack API along
with the raw Slack Message object. This was useful for middleware and
scripts which reformatted Slack text in different ways than hubot-slack's
own default formatter.

For example, a middleware might try to detect URLs that were
autoexpanded by the Slack API and restore them to their original format;
this is only possible if the raw text is included, so the unparsed URLs
can be read.