implements SlackClient#onMessage() and deprecates SlackClient#on()
in order to implement tests for `user.list` pagination, the
eventemitter tests (which were incorrect) had to be removed. the
`.on()` method was stubbed such that it invoked the callback
synchronously on handler registration. instead of stubbing this
method, this commit removes the event emitter responsiblity from
the SlackClient class and just delegates this responsibility onto
the RTM client (which inherits this from EventEmitter3). now its
not required to test that functionality since its covered in
dependencies.

once these tests were removed, an `onMessage()` method was added
in order to provide the message handler wrapping functionality
that was previously implemented in the removed `on()` method. this
method takes a purpose-specific callback for 'message' events, which
is the only event type that SlackClient provided any custom handling
for.

new tests were added to cover the changes. the coverage diff may be
lower than necessary, and more tests may need to be added.