Add flag to disable loading all users on hubot startup
For a team with a large number of users, auto-loading all users uses a
lot of space in the Hubot brain.  This commit adds an environment flag:
NO_LOAD_ALL_USERS.  When set, the loadUsers event will NOT trigger
userChange events for all users, instead it will be skipped and users
will only be loaded on-demand, when the bot hears them say something.

Fixes #467