Upgrade @slack/client to 4.2.0 to address CVE-2018-3728
The 3.x branch of @slack/client has been abandoned and contains a locked
reference to an older version of request which contains a vulnerable
version of hoek. The only way forward is to move to the 4.x branch of
@slack/client.

The migration guide can be found here:
https://github.com/slackapi/node-slack-sdk/wiki/Migration-Guide-for-v4

In order to get the suite to pass I had to s/RtmClient/RTMClient/g
across the app. I also had to modify 3 tests to mimic API changes
upstream, but once done all tests pass and no source code modifications
other than the capitalization of RTMClient were necessary.

This could use lots of eyes.