Fix unexpected newline after super
So I know that the project has specifically pinned coffeescript 1.x in
its package.json but I'm trying to use coffeescript 2.x in my project
because it has other things I need (namely async/await) and the bot
crashes on start referencing this line from bot.coffee:

  [stdin]:20:10: error: unexpected newline
      super
           ^.

I supposed that this keyword has now become ambiguous without brackets
but to make sure, I dug up the breaking changes list for the
coffeescript 2.0.0 release and found this:

> Due to a syntax clash with super with accessors, "bare" super (the
> keyword super without parentheses) no longer compiles to a super call
> forwarding all arguments.
> [...] if you know that the parent function doesn't require arguments,
> just call `super()`.

Source: https://coffeescript.org/#breaking-changes-super-extends

After adding the brackets the coffeelint tool no longer spits out that
error and when running it it "works on my machine".  Tests are still
passing and also I can't it imagine it having any adverse effects on
users of older versions of coffeescript because the parentheses were
optional so far.